+++ Angular-Front-End/app-front-end/src/app/components/communities-list/communities-list.component.ts
@@ -22,7 +22,7 @@ export class CommunitiesListComponent implements OnInit {
 
   getCommunityList(): void {
     this.communitiesService.getAll().subscribe((communitiesList) => {
-      this.communities = communitiesList;
+      this.communities = communitiesList.second.list;
     })
   }
 }