+++ Angular-Front-End/app-front-end/src/app/models/user.model.ts
@@ -1,0 +1,7 @@
+import { Identifiable } from './identifiable';
+
+export interface UserModel extends Identifiable{
+    userName: String;
+    email: String;
+    password: String;
+}