+++ Angular-Front-End/app-front-end/src/app/services/controllers/communities-controller.service.ts
@@ -14,7 +14,7 @@ export class CommunitiesService extends TemplateControllerService<CommunityModel
     super(http);
   }
 
-  private communitiesUrl = '/communityCatalogApi/communityCatalog/communities';
+  private communitiesUrl = '/communityCatalogApi/communities';
   private getCommunityById = '/communityCatalogApi/community/';
 
   // Abstract class methods