+++ Angular-Front-End/app-front-end/src/environments/environment.prod.ts
@@ -1,3 +1,13 @@
 export const environment = {
-  production: true
+  production: true,
+  firebaseConnection: {
+    apiKey: "AIzaSyAkNJU2HU3CT9S7UnRlDFl4YYNYZmQB7xw",
+    authDomain: "fcul-app-testing.firebaseapp.com",
+    databaseURL: "https://fcul-app-testing.firebaseio.com",
+    projectId: "fcul-app-testing",
+    storageBucket: "fcul-app-testing.appspot.com",
+    messagingSenderId: "801529151844",
+    appId: "1:801529151844:web:1d956f6ae5736402f67a71",
+    measurementId: "G-QC7778J7N7"
+  }
 };