+++ Database/SQLTables.sql
@@ -3,100 +3,131 @@
 -- ------------------------------------------------------
 -- Server version	5.7.29-0ubuntu0.18.04.1
 
+
+-- ------------------------------------------------------
+-- ---------------------Database-------------------------
+-- ------------------------------------------------------
+USE RolesUtilCommunities;
 -- ------------------------------------------------------
 -- ---------------------Drop Tables----------------------
 -- ------------------------------------------------------
-DROP TABLE IF EXISTS Events CASCADE;
 DROP TABLE IF EXISTS RolesUsersCommunities CASCADE;
+DROP TABLE IF EXISTS Events CASCADE;
 DROP TABLE IF EXISTS Users CASCADE;
-DROP TABLE IF EXISTS Communities CASCADE;
 DROP TABLE IF EXISTS Roles CASCADE;
+DROP TABLE IF EXISTS Communities CASCADE;
 
 -- ------------------------------------------------------
 -- --------------------Create Tables---------------------
 -- ------------------------------------------------------
 
 CREATE TABLE Users (
-	uID integer PRIMARY KEY auto_increment, 
-	uName varchar(20) not null
+	`uID` integer auto_increment, 
+	`uName` varchar(50) NOT NULL,
+	`uEmail` varchar(50) NOT NULL,
+	`uPassword` varchar(20) NOT NULL,
+	PRIMARY KEY (`uID`),
+	UNIQUE KEY `uName` (`uName`)
 );
 
 CREATE TABLE Communities (
-	cID integer PRIMARY KEY auto_increment, 
-	cName varchar(20) UNIQUE not null
+	`cID` integer auto_increment, 
+	`cName` varchar(20) NOT NULL,
+	`description` varchar(500) NOT NULL,
+	PRIMARY KEY (`cID`),
+	UNIQUE KEY `cName` (`cName`)
 );
 
--- ---------TODO---------
 CREATE TABLE Roles (
-	rID integer PRIMARY KEY auto_increment, 
-	rName varchar(20) UNIQUE not null,
-	cID varchar(20),
-	aLevel integer not null
-	CONSTRAINT fk_cID FOREIGN KEY (cID) REFERENCES Communities(cID) ON DELETE CASCADE ON UPDATE CASCADE
+	`id` int(11) NOT NULL AUTO_INCREMENT,
+	`rName` varchar(20) NOT NULL,
+	`authLevel` int(1) NOT NULL,
+	`cID` int(11) NOT NULL,
+	PRIMARY KEY (`id`),
+	UNIQUE KEY `rName` (`rName`),
+	FOREIGN KEY (cID) REFERENCES Communities(cID) ON DELETE CASCADE ON UPDATE CASCADE
 );
--- -----------------------
 
 CREATE TABLE Events (
-	id integer PRIMARY KEY auto_increment, 
-	eName varchar(50) UNIQUE not null, 
-	cID integer, 
-	cName varchar(20),
-	dStart DATE not null, 
-	dEnd DATE not null, 
-	CONSTRAINT fk_cID FOREIGN KEY (cID) REFERENCES Communities(cID) ON DELETE CASCADE ON UPDATE CASCADE
+	`id` int(11) NOT NULL AUTO_INCREMENT,
+	`eName` varchar(30) NOT NULL,
+	`start` date NOT NULL,
+	`end` date NOT NULL,
+	`cID` int(11) NOT NULL,
+	`cName` varchar(20) NOT NULL,
+	PRIMARY KEY (`id`),
+	UNIQUE KEY `eName` (`eName`),
+	FOREIGN KEY (cID) REFERENCES Communities(cID) ON DELETE CASCADE ON UPDATE CASCADE
 );
 
 CREATE TABLE RolesUsersCommunities (
-	id integer PRIMARY KEY auto_increment, 
-	uID integer, 
-	cID2 integer, 
-	rID integer, 
-	dStart DATE not null, 
-	dEnd DATE not null,
-	CONSTRAINT fk_uID FOREIGN KEY (uID) REFERENCES Users(uID) ON DELETE CASCADE ON UPDATE CASCADE,
-	CONSTRAINT fk_cID2 FOREIGN KEY (cID2) REFERENCES Communities(cID) ON DELETE CASCADE ON UPDATE CASCADE,
-	CONSTRAINT fk_rID FOREIGN KEY (rID) REFERENCES Roles(rID) ON DELETE CASCADE ON UPDATE CASCADE
+	`id` int(11) NOT NULL AUTO_INCREMENT,
+	`uID` int(11) NOT NULL,
+	`cID` int(11) NOT NULL,
+	`rID` int(11) NOT NULL,
+	`dStart` date NOT NULL,
+	`dEnd` date NOT NULL,
+	PRIMARY KEY (`id`),
+	UNIQUE KEY `ruc` (`uID`,`cID`,`rID`),
+	FOREIGN KEY (uID) REFERENCES Users(uID) ON DELETE CASCADE ON UPDATE CASCADE,
+	FOREIGN KEY (cID) REFERENCES Communities(cID) ON DELETE CASCADE ON UPDATE CASCADE,
+	FOREIGN KEY (rID) REFERENCES Roles(id) ON DELETE CASCADE ON UPDATE CASCADE
 );
 
 -- ------------------------------------------------------
 -- -------------------Insert Values----------------------
 -- ------------------------------------------------------
 
--- 
--- All users are subscribed to a community, the ALL_USER_COMMUNITY
-LOCK TABLES Users WRITE;
-INSERT INTO Communities VALUES(1, 'ALL_USER_COMMUNITY');
+-- --------------------Communities-----------------------
+LOCK TABLES Communities WRITE;
+INSERT INTO Communities VALUES(1, 'ADMIN_COMMUNITY', 'All Admins belong to this community');
+INSERT INTO Communities VALUES(2, 'MAINTAINER_COMMUNITY', 'All Maintainers belong to this community');
+INSERT INTO Communities VALUES(3, 'ALL_USER_COMMUNITY', 'All Users belong to this community');
+
+-- Test Values
+INSERT INTO Communities VALUES (4,'ASC', 'This is meant for all the students that  are participating in the
+	course of ASC'),
+	(5,'LASIGE', 'Computer Science and Engineering research unit at the Department of Informatics, 
+	Faculty of Sciences, University of Lisboa.'),
+	(6,'SD', 'This is meant for all the students that  are participating in the
+	course of SD'),
+	(7,'SO', 'This is meant for all the students that  are participating in the
+	course of SO');
 UNLOCK TABLES;
 
+
+-- ----------------------Roles------------------------
 -- There is a default role, which all users get when they enter the system. All users
 -- are given a role on the ALL_USER_COMMUNITY, which they belong to.
 -- Roles go from authLevel 1 to 3, 1 being the authLevel with the least privileges
 Lock Tables Roles WRITE;
-INSERT INTO Roles Values(1, 'DEFAULT_ROLE', 1, 1);
-INSERT INTO Roles Values(1, 'MAINTAINER_ROLE', 1, 2);
-INSERT INTO Roles Values(2, 'ADMIN_ROLE', 1, 3);
+INSERT INTO Roles Values(1, 'ADMIN_ROLE', 1, 3);
+INSERT INTO Roles Values(2, 'MAINTAINER_ROLE', 1, 2);
+INSERT INTO Roles Values(3, 'DEFAULT_USER_ROLE', 1, 1);
+
+
+
+-- Test Values
+INSERT INTO Roles VALUES (4, 'Hero-Admin', 3, 1);
 UNLOCK TABLES;
 
--- ------------------------------------------------------
--- --------------------- Test Values --------------------
--- ------------------------------------------------------
-LOCK TABLES Communities WRITE;
-INSERT INTO Communities VALUES (2,'ASC'),(3,'LASIGE'),(4,'SD'),(5,'SO');
-UNLOCK TABLES;
-
+-- ----------------------Users------------------------
 LOCK TABLES Users WRITE;
-INSERT INTO Users VALUES (2,'Manuel'),(1,'Pedro');
+-- Test Values
+INSERT INTO Users VALUES (2,'Manuel','manuel@gmail.com', 'manuelPassword'),(1,'Pedro','pedro@outlook.pt', 'pedroPassword');
 UNLOCK TABLES;
 
+-- --------------------Events-------------------------
 LOCK TABLES Events WRITE;
-INSERT INTO Events VALUES (1,'Reunião',1,'ASC', '2020-02-24','2020-02-25'),(2,'Almoço_no_LASIGE',2,'LASIGE','2020-03-14','2020-03-14');
+-- Test Values
+INSERT INTO Events VALUES (1,'Reunião','2020-02-24','2020-02-25',1,'ASC'),
+			  (2,'Almoço_no_LASIGE','2020-03-14','2020-03-14',2,'LASIGE');
 UNLOCK TABLES;
 
-LOCK TABLES Roles WRITE;
-INSERT INTO Roles VALUES (1, 'Hero', 5);
-UNLOCK TABLES;
-
+-- --------------RolesUsersCommunities----------------
 LOCK TABLES RolesUsersCommunities WRITE;
-INSERT INTO RolesUsersCommunities VALUES (1,1,1,1,'2020-01-04','2021-01-04'),(2,2,2,1,'2020-01-04','2021-01-04');
+-- Test Values
+INSERT INTO RolesUsersCommunities VALUES (1,1,4,3,'2020-01-04','2021-01-04'),
+					 (2,2,5,3,'2020-01-04','2021-01-04');
 UNLOCK TABLES;