+++ Services/ComunityCatalogService/src/main/java/io/App/CommunityService/communityComponent/Community.java
@@ -4,10 +4,12 @@ public class Community {
 
 	private int id;
 	private String name;
+	private String description;
 
-	public Community(int id, String name) {
+	public Community(int id, String name, String description) {
 		this.id = id;
 		this.name = name;
+		this.description = description;
 	}
 
 	public Community(String name) {
@@ -26,6 +28,10 @@ public class Community {
 	public String getName() {
 		return name;
 	}
+	
+	public String getDescription() {
+		return description;
+	}
 
 	public void setId(int id) {
 		this.id = id;
@@ -35,4 +41,7 @@ public class Community {
 		this.name = name;
 	}
 
+	public void setDescription(String description) {
+		this.description = description;
+	}
 }