+++ Services/ComunityCatalogService/src/main/java/io/App/CommunityService/dto/Pair.java
@@ -1,4 +1,4 @@
-package io.App.CommunityService.communityComponent;
+package io.App.CommunityService.dto;
 
 import java.io.Serializable;